tests/test_files_generator/requirements.txt
python-docx>=0.8.11
openpyxl>=3.1.2
Pillow>=10.0.0
